Break out unit tests from e2e
