+++ test/unit/db/level.test.ts
@@ -1,73 +1,79 @@
 import { assert } from "chai";
-import {
-  serialize,
-  hashTreeRoot,
-} from "@chainsafe/ssz";
+import leveldown from "leveldown";
+import promisify from "promisify-es6";
+import {hashTreeRoot, serialize,} from "@chainsafe/ssz";
 
-import {
-  BeaconBlock,
-  BeaconState,
-} from "../../src/types";
+import {BeaconBlock, BeaconState,} from "../../../src/types";
 
-import {PouchDb} from "../../src/db";
+import {LevelDB} from "../../../src/db";
 
-import { generateState } from "../utils/state";
-import { generateEmptyBlock } from "../utils/block";
-import { generateEmptyAttestation } from "../utils/attestation";
-import {generateEmptyVoluntaryExit} from "../utils/voluntaryExits";
+import {generateState} from "../../utils/state";
+import {generateEmptyBlock} from "../../utils/block";
+import {generateEmptyAttestation} from "../../utils/attestation";
+import {generateEmptyVoluntaryExit} from "../../utils/voluntaryExits";
+import level from "level";
 
-describe("PouchDB", () => {
-  const db = new PouchDb({name: 'testdb'});
-
-  after(async () => {
-    await db.clean();
+describe("LevelDB", () => {
+  const dbLocation = "./.__testdb";
+  const testDb = level(
+    dbLocation, {
+      keyEncoding: 'binary',
+      valueEncoding: 'binary',
+    });
+  const db = new LevelDB({db: testDb});
+  before(async () => {
+    await db.start();
   });
-
+  after(async () => {
+    await db.stop();
+    await promisify(leveldown.destroy)(dbLocation, function () {
+    })
+  });
   it("should correctly get and set the state", async () => {
     const testState = generateState();
     await db.setState(testState);
     const actualState = await db.getState();
     assert.deepEqual(serialize(actualState, BeaconState), serialize(testState, BeaconState));
-  })
+  });
   it("should correctly get and set the finalized state", async () => {
     const testState = generateState();
     await db.setFinalizedState(testState);
     const actualState = await db.getFinalizedState();
     assert.deepEqual(serialize(actualState, BeaconState), serialize(testState, BeaconState));
-  })
+  });
   it("should correctly get and set a block", async () => {
     const testBlock = generateEmptyBlock();
     const testBlockRoot = hashTreeRoot(testBlock, BeaconBlock);
     await db.setBlock(testBlock);
     const actualBlock = await db.getBlock(testBlockRoot);
     assert.deepEqual(serialize(actualBlock, BeaconBlock), serialize(testBlock, BeaconBlock));
-  })
+  });
   it("should correctly set the chain head", async () => {
-      const testState = generateState();
-      const testBlock = generateEmptyBlock();
-      const slot = 5;
-      testBlock.slot = slot;
-      const testBlockRoot = hashTreeRoot(testBlock, BeaconBlock);
-      await db.setBlock(testBlock);
-      await db.setChainHead(testState, testBlock);
-      const actualBlock = await db.getBlockBySlot(slot);
-      assert.deepEqual(serialize(actualBlock, BeaconBlock), serialize(testBlock, BeaconBlock));
-  })
+    const testState = generateState();
+    const testBlock = generateEmptyBlock();
+    const slot = 5;
+    testBlock.slot = slot;
+    const testBlockRoot = hashTreeRoot(testBlock, BeaconBlock);
+    await db.setBlock(testBlock);
+    await db.setChainHead(testState, testBlock);
+    const actualBlock = await db.getBlockBySlot(slot);
+    assert.deepEqual(serialize(actualBlock, BeaconBlock), serialize(testBlock, BeaconBlock));
+  });
   it("should correctly set, get, delete attestations", async () => {
     const testAttestations = Array.from({length: 64}, (_, i) => {
       const a = generateEmptyAttestation();
       a.aggregationBitfield = Buffer.from(i.toString());
       return a;
-    })
+    });
     for (const a of testAttestations) {
       await db.setAttestation(a);
     }
     const actualAttestations = await db.getAttestations();
     assert.equal(actualAttestations.length, testAttestations.length);
-    await db.deleteAttestations(actualAttestations);
+    await db.deleteAttestations(testAttestations);
     const noAttestations = await db.getAttestations();
     assert.equal(noAttestations.length, 0);
-  })
+  });
 
   it("should correctly set, get, delete voluntary exits", async () => {
       const testVoluntaryExits = Array.from({length: 10}, (_, i) => {
