+++ test/unit/eth1/ethers.test.ts
@@ -4,8 +4,8 @@ import { ethers, Event } from "ethers";
 import ganache from "ganache-core";
 import sinon from "sinon";
 
-import { EthersEth1Notifier } from "../../src/eth1";
-import defaults from "../../src/eth1/defaults";
+import { EthersEth1Notifier } from "../../../src/eth1";
+import defaults from "../../../src/eth1/defaults";
 import promisify from "promisify-es6";
 
 describe("Eth1Notifier", () => {
