+++ test/unit/rpc/jsonRpcOverWs.test.ts
@@ -1,9 +1,9 @@
 import { assert } from "chai";
 import * as jsonRpc from "noice-json-rpc";
 import Websocket from "ws";
-import {MockAPI, JSONRPC, API, WSServer} from "../../src/rpc";
-import { generateEmptyBlock } from "../utils/block";
-import { generateEmptyAttestation } from "../utils/attestation";
+import {MockAPI, JSONRPC, API, WSServer} from "../../../src/rpc";
+import { generateEmptyBlock } from "../../utils/block";
+import { generateEmptyAttestation } from "../../utils/attestation";
 
 describe("Json RPC over WS", () => {
   const rpc = new JSONRPC({}, {transport: new WSServer({port: 32420}), api: new MockAPI()});
